1 Перемещены типы counter и gauge в хранилище (в памяти == файловое)
